+++ ideas.md
@@ -1,0 +1,2 @@
+1. Add caching to API requests: https://pypi.org/project/requests-cache/
+2. add support for skipping files already downloaded
