+++ sec_edgar_downloader/_version.py
@@ -1,1 +1,1 @@
-__version__ = "4.3.0"
+__version__ = "5.0.0"
